@@ -2,11 +2,11 @@
   "name": "brandprolab",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
     "format": "prettier --write .",
     "lint:fix": "next lint --fix",
     "prepare": "husky install",
@@ -23,6 +23,7 @@
     "@storybook/addon-links": "^8.5.0",
     "@storybook/addon-viewport": "^8.5.0",
     "classnames": "^2.5.1",
+    "esbuild": "^0.25.0",
     "next": "15.1.2",
     "next-sanity": "^9.8.31",
     "react": "^19.0.0",
@@ -33,6 +34,7 @@
   "devDependencies": {
     "@chromatic-com/storybook": "^3.2.4",
     "@eslint/eslintrc": "^3",
+    "@next/eslint-plugin-next": "^15.1.6",
     "@storybook/addon-a11y": "^8.5.0",
     "@storybook/addon-essentials": "^8.5.0",
     "@storybook/addon-interactions": "^8.5.0",
@@ -50,19 +52,26 @@
     "@types/node": "^20",
     "@types/react": "^19.0.2",
     "@types/react-dom": "^19.0.2",
-    "@typescript-eslint/parser": "^8.18.1",
+    "@typescript-eslint/eslint-plugin": "^8.23.0",
+    "@typescript-eslint/parser": "^8.23.0",
     "autoprefixer": "^10.4.20",
     "commitizen": "^4.3.1",
     "css-loader": "^7.1.2",
     "cz-conventional-changelog": "^3.3.0",
-    "eslint": "^9",
-    "eslint-config-next": "15.1.2",
+    "depcheck": "^1.4.7",
+    "eslint": "^9.20.0",
+    "eslint-config-next": "^15.1.6",
     "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-eslint-comments": "^3.2.0",
     "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
     "eslint-plugin-playwright": "^2.1.0",
-    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-prettier": "^5.2.3",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-hooks": "^5.1.0",
     "eslint-plugin-simple-import-sort": "^12.1.1",
     "eslint-plugin-storybook": "^0.11.2",
+    "eslint-plugin-tailwindcss": "^3.18.0",
     "eslint-plugin-unicorn": "^56.0.1",
     "happy-dom": "^15.11.7",
     "husky": "^8.0.0",
@@ -72,8 +81,8 @@
     "plop": "^4.0.1",
     "postcss": "^8.4.31",
     "postcss-loader": "^8.1.1",
-    "prettier": "^3.4.2",
-    "prettier-plugin-tailwindcss": "^0.6.9",
+    "prettier": "^3.5.0",
+    "prettier-plugin-tailwindcss": "^0.6.11",
     "storybook": "^8.5.0",
     "style-loader": "^4.0.0",
     "tailwindcss": "^3.4.17",
