@@ -1,10 +1,12 @@
 import { render, screen } from '@testing-library/react';
+
 import Container from './Container';
+
 import '@testing-library/jest-dom';
 import {
   FULL_WIDTH_CONTAINER_TEST_ID,
   LIMITED_WIDTH_CONTAINER_TEST_ID,
-} from '@/lib/constants';
+} from '@/lib/testIDs';
 
 describe('Container component', () => {
   it('renders children inside the full-width container', () => {
