@@ -1,7 +1,9 @@
 import { render, screen } from '@testing-library/react';
+
 import UnderlinedText from './UnderlinedText';
+
 import '@testing-library/jest-dom';
-import { UNDERLINED_TEXT_TEST_ID } from '@/lib/constants';
+import { UNDERLINED_TEXT_TEST_ID } from '@/lib/testIDs';
 
 describe('UnderlinedText component', () => {
   const defaultProps = {
