@@ -1,10 +1,11 @@
+import { UNDERLINED_TEXT_TEST_ID } from '@/lib/testIDs';
+
 import { UnderlinedTextProps } from './types';
-import { UNDERLINED_TEXT_TEST_ID } from '@/lib/constants';
 
 const UnderlinedText = ({ children }: UnderlinedTextProps) => {
   return (
     <span
-      className="relative inline-block cursor-pointer text-16 font-bold leading-none text-teal-dark underline"
+      className="relative inline-block cursor-pointer text-base font-bold leading-none text-teal-dark underline"
       data-testid={UNDERLINED_TEXT_TEST_ID}
     >
       {children}
