@@ -1,10 +1,9 @@
-import { getIconComponentStyles } from '@/utils';
 import { useMemo } from 'react';
 
-export const useIconStyles = (
-  size: 's' | 'm' | 'l',
-  isIconCentered: boolean,
-) => {
+import { getIconComponentStyles } from '@/utils';
+import { Size } from '@/lib/types';
+
+export const useIconStyles = (size: Size, isIconCentered: boolean) => {
   return useMemo(
     () => getIconComponentStyles(size, isIconCentered),
     [size, isIconCentered],
