@@ -1,11 +1,10 @@
 import { client } from '@/sanity/lib/client';
+import { throwError } from './throwError';
 
 export async function fetchDataFromSanity<T>(query: string): Promise<T> {
   try {
-    const data = await client.fetch<T>(query);
-    return data;
+    return await client.fetch<T>(query);
   } catch (error) {
-    console.error(`Error fetching data:`, error);
-    throw new Error(`Failed to fetch data.`);
+    throwError(`Failed to fetch data.`);
   }
 }
