@@ -1,8 +1,7 @@
-import { BACKGROUND_CIRCLE_SIZES } from '@/lib/constants';
+import { BACKGROUND_CIRCLE_SIZES } from '@/styles/constants';
+import { getSizeFromConstants } from './getSizeFromConstants';
+import { Size, Screen } from '@/lib/types';
 
-export function getCircleSize(
-  circleSize: 's' | 'm' | 'l',
-  screen: 'mobile' | 'tablet' | 'desktop',
-) {
-  return BACKGROUND_CIRCLE_SIZES[circleSize][screen];
+export function getCircleSize(circleSize: Size, screen: Screen) {
+  return getSizeFromConstants(BACKGROUND_CIRCLE_SIZES, circleSize, screen);
 }
