@@ -1,13 +1,13 @@
 import {
   BACKGROUND_CIRCLE_SIZES,
   ICON_OFFSET_COEFFICIENT,
-} from '@/lib/constants';
+} from '@/styles/constants';
+import { getSizeFromConstants } from './getSizeFromConstants';
+import { Screen, Size } from '@/lib/types';
 
-export function getIconOffset(
-  circleSize: 's' | 'm' | 'l',
-  screen: 'mobile' | 'tablet' | 'desktop',
-) {
+export function getIconOffset(circleSize: Size, screen: Screen) {
   return Math.ceil(
-    BACKGROUND_CIRCLE_SIZES[circleSize][screen] / ICON_OFFSET_COEFFICIENT,
+    getSizeFromConstants(BACKGROUND_CIRCLE_SIZES, circleSize, screen) /
+      ICON_OFFSET_COEFFICIENT,
   );
 }
